requirements.txt
# Django & REST API
Django>=5.0,<5.1
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0

# Database (PostgreSQL)
psycopg2-binary>=2.9,<3.0

# Celery & Redis
celery>=5.4,<6.0
redis>=5.0,<6.0

# Video Processing
ffmpeg-python>=0.2.0
openai>=1.0.0,<2.0
openai-whisper>=20231117
yt-dlp>=2023.12.30
numpy>=1.24,<2.0
moviepy>=1.0.3,<2.0

# Audio processing for AI moment detection
librosa>=0.10.0,<0.11.0
soundfile>=0.12.0,<0.13.0

# Storage & File Handling
boto3>=1.34,<2.0
Pillow>=9.0,<10.0

# Production Server
gunicorn>=21.2,<22.0

# Environment Variables
python-dotenv>=1.0,<2.0

# Payment Processing
stripe>=12.0.0,<13.0

# PyTorch CPU-only (compatible versions)
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torch==2.1.2+cpu
torchaudio==2.1.2+cpu

# Additional dependencies that might be needed
imageio[ffmpeg]>=2.31.1
imageio-ffmpeg>=0.4.8
decorator>=4.4.0